chore: simplify graylog tests
The TCP graylog tests are one of the most flaky tests we have. The
current scenario attempts to send two messages, then close the
connection, fail to send a third, and then correctly send a fourth.

Because these failures are commonly only seen in CI, I want to see
if this simplified test still results in failures. My hunch is that
the simulated 3rd packet failure has a race condition in it causing
some sort of failure/timeout to get hit.

This also runs circleci only run mac tests, since that is where most
failures are seen.